backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.5.1
python-socketio==5.9.0
python-engineio==4.7.0
torch==2.0.0
torchvision==0.15.0
numpy==1.24.4
opencv-python==4.8.0.76
scipy==1.10.0
pandas==2.0.0
mne==1.3.0
scikit-learn==1.2.0
matplotlib==3.7.0
seaborn==0.12.0
python-dotenv==1.0.0
Werkzeug==2.3.0
requests

 